Experiment with data control for processing block settings
